ENH:  Add support for adding tracks in animation gui.       New class pqAnimatableProxyComboBox so pqAnimationPanel and         pqAnimationViewWidget can share some code. STYLE:  Some clean up too.

+++ Qt/Widgets/pqAnimationWidget.cxx
@@ -54,13 +54,13 @@ pqAnimationWidget::pqAnimationWidget(QWidget* p)
   this->Model = new pqAnimationModel(this->View);
   this->View->setScene(this->Model);
 
-  this->AddRemoveHeader = new QHeaderView(Qt::Vertical, this);
-  this->AddRemoveHeader->setClickable(true);
-  this->AddRemoveHeader->setSizePolicy(QSizePolicy::Minimum,
+  this->CreateDeleteHeader = new QHeaderView(Qt::Vertical, this);
+  this->CreateDeleteHeader->setClickable(true);
+  this->CreateDeleteHeader->setSizePolicy(QSizePolicy::Minimum,
                               QSizePolicy::MinimumExpanding);
 
-  this->AddRemoveHeader->setResizeMode(QHeaderView::Fixed);
-  this->AddRemoveHeader->setModel(&this->AddRemoveModel);
+  this->CreateDeleteHeader->setResizeMode(QHeaderView::Fixed);
+  this->CreateDeleteHeader->setModel(&this->CreateDeleteModel);
 
   this->Header = new QHeaderView(Qt::Vertical, this);
   this->Header->setObjectName("TrackHeader");
@@ -71,7 +71,10 @@ pqAnimationWidget::pqAnimationWidget(QWidget* p)
   this->Header->setResizeMode(QHeaderView::Fixed);
   this->Header->setMinimumSectionSize(0);
   this->Header->setModel(this->Model->header());
-  this->Model->setRowHeight(this->Header->sectionSize(0));
+  this->Model->setRowHeight(this->Header->defaultSectionSize());
+  
+  this->CreateDeleteWidget = new QWidget(this);
+
   QObject::connect(this->Header->model(),
                    SIGNAL(rowsInserted(QModelIndex,int,int)),
                    this, SLOT(updateSizes()));
@@ -87,7 +90,7 @@ pqAnimationWidget::pqAnimationWidget(QWidget* p)
   QObject::connect(this->Model,
                    SIGNAL(trackSelected(pqAnimationTrack*)),
                    this, SIGNAL(trackSelected(pqAnimationTrack*)));
-  QObject::connect(this->AddRemoveHeader,
+  QObject::connect(this->CreateDeleteHeader,
                    SIGNAL(sectionClicked(int)),
                    this, SLOT(headerDeleteClicked(int)));
 }
@@ -103,30 +106,34 @@ pqAnimationModel* pqAnimationWidget::animationModel() const
 
 QHeaderView* pqAnimationWidget::createDeleteHeader() const
 {
-  return this->AddRemoveHeader;
+  return this->CreateDeleteHeader;
+}
+
+QWidget* pqAnimationWidget::createDeleteWidget() const
+{
+  return this->CreateDeleteWidget;
 }
 
 void pqAnimationWidget::updateSizes()
 {
-  this->AddRemoveModel.clear();
-  this->AddRemoveModel.insertRow(0);
-  this->AddRemoveModel.setHeaderData(0, Qt::Vertical, QVariant(), Qt::DisplayRole);
+  this->CreateDeleteModel.clear();
+  this->CreateDeleteModel.insertRow(0);
+  this->CreateDeleteModel.setHeaderData(0, Qt::Vertical, QVariant(), Qt::DisplayRole);
 
   int num = this->Model->count();
 
   for(int i=0; i<num; i++)
     {
-    this->AddRemoveModel.insertRow(i+1);
+    this->CreateDeleteModel.insertRow(i+1);
     if(this->Model->track(i)->isDeletable())
       {
-      this->AddRemoveModel.setHeaderData(i+1, Qt::Vertical,
+      this->CreateDeleteModel.setHeaderData(i+1, Qt::Vertical,
         QPixmap(":/QtWidgets/Icons/pqDelete16.png"), Qt::DecorationRole);
       }
-    this->AddRemoveModel.setHeaderData(i+1, Qt::Vertical, QVariant(), Qt::DisplayRole);
+    this->CreateDeleteModel.setHeaderData(i+1, Qt::Vertical, QVariant(), Qt::DisplayRole);
     }
-  
-  this->AddRemoveModel.insertRow(this->Header->count());
-  this->AddRemoveModel.setHeaderData(this->Header->count(), Qt::Vertical,
+  this->CreateDeleteModel.insertRow(this->Header->count());
+  this->CreateDeleteModel.setHeaderData(this->Header->count(), Qt::Vertical,
     QPixmap(":/QtWidgets/Icons/pqPlus16.png"), Qt::DecorationRole);
   
   this->updateGeometries();
@@ -145,24 +152,30 @@ void pqAnimationWidget::updateGeometries()
 {
   int width1 = 0;
   int width2 = 0;
-  if(!this->AddRemoveHeader->isHidden())
+  int height1 = 0;
+  int height2 = 0;
+  if(!this->CreateDeleteHeader->isHidden())
     {
-    int tmp = qMax(this->AddRemoveHeader->minimumWidth(),
-                 this->AddRemoveHeader->sizeHint().width());
-    width1 = qMin(tmp, this->AddRemoveHeader->maximumWidth());
+    int tmp = qMax(this->CreateDeleteHeader->minimumWidth(),
+                 this->CreateDeleteHeader->sizeHint().width());
+    width1 = qMin(tmp, this->CreateDeleteHeader->maximumWidth());
+    height1 = this->CreateDeleteHeader->defaultSectionSize() *
+              this->CreateDeleteHeader->count();
     }
   if(!this->Header->isHidden())
     {
     int tmp = qMax(this->Header->minimumWidth(),
                  this->Header->sizeHint().width());
     width2 = qMin(tmp, this->Header->maximumWidth());
+    height2 = this->Header->defaultSectionSize() *
+              this->Header->count();
     }
 
   this->setViewportMargins(width1 + width2, 0, 0, 0);
 
   QRect vg = this->contentsRect();
-  this->AddRemoveHeader->setGeometry(vg.left(), vg.top(), width1, vg.height());
-  this->Header->setGeometry(vg.left() + width1, vg.top(), width2, vg.height());
+  this->CreateDeleteHeader->setGeometry(vg.left(), vg.top(), width1, height1);
+  this->Header->setGeometry(vg.left() + width1, vg.top(), width2, height2);
 
   this->updateScrollBars();
 }
@@ -171,7 +184,7 @@ void pqAnimationWidget::scrollContentsBy(int dx, int dy)
 {
   if(dy)
     {
-    this->AddRemoveHeader->setOffset(this->verticalScrollBar()->value());
+    this->CreateDeleteHeader->setOffset(this->verticalScrollBar()->value());
     this->Header->setOffset(this->verticalScrollBar()->value());
     }
   this->updateWidgetPosition();
@@ -181,17 +194,25 @@ void pqAnimationWidget::scrollContentsBy(int dx, int dy)
 void pqAnimationWidget::updateScrollBars()
 {
   int h = this->View->sizeHint().height();
-  if(this->AddRemoveHeader->isVisible())
+  int extraw = 0;
+  int viewh = h;
+  if(this->CreateDeleteHeader->isVisible())
     {
-    h = qMax(h, this->AddRemoveHeader->length());
+    h = qMax(h, this->CreateDeleteHeader->length());
     }
   if(this->Header->isVisible())
     {
     h = qMax(h, this->Header->length());
+    extraw = this->Header->width();
+    viewh = h;
     }
   
   QSize vsize = this->viewport()->size();
   this->View->resize(vsize.width(), h);
+  this->CreateDeleteWidget->resize(vsize.width()+extraw,
+                                   this->Header->defaultSectionSize());
+
+  this->updateWidgetPosition();
 
   this->verticalScrollBar()->setPageStep(vsize.height());
   this->verticalScrollBar()->setRange(0, h-vsize.height());
@@ -199,14 +220,23 @@ void pqAnimationWidget::updateScrollBars()
 
 void pqAnimationWidget::updateWidgetPosition()
 {
-  this->View->move(0, -this->verticalScrollBar()->value());
+  int s = this->verticalScrollBar()->value();
+  this->View->move(0, -s);
+  int m = this->View->frameGeometry().bottom()+1-
+          this->CreateDeleteHeader->defaultSectionSize();
+  int w = 0;
+  if(this->CreateDeleteHeader->isVisible())
+    {
+    w = this->CreateDeleteHeader->frameGeometry().right()+1;
+    }
+  this->CreateDeleteWidget->move(w, m);
 }
 
 bool pqAnimationWidget::event(QEvent* e)
 {
   if(e->type() == QEvent::FontChange)
     {
-    this->Model->setRowHeight(this->Header->sectionSize(0));
+    this->Model->setRowHeight(this->Header->defaultSectionSize());
     }
   if(e->type() == QEvent::Show)
     {
@@ -225,10 +255,17 @@ void pqAnimationWidget::headerDeleteClicked(int which)
 {
   if(which > 0)
     {
-    pqAnimationTrack* t = this->Model->track(which-1);
-    if(t && t->isDeletable())
+    if(which == this->CreateDeleteHeader->count() - 1)
       {
-      emit this->deleteTrackClicked(t);
+      emit this->createTrackClicked();
+      }
+    else
+      {
+      pqAnimationTrack* t = this->Model->track(which-1);
+      if(t && t->isDeletable())
+        {
+        emit this->deleteTrackClicked(t);
+        }
       }
     }
 }
